refactor:#265 home 리팩토링
