+++ src/app/(home)/_components/home-calendar/home-calendar-month.tsx
@@ -1,31 +1,32 @@
+/* eslint-disable react-hooks/exhaustive-deps */
 import { Carousel, CarouselContent, CarouselItem } from '@/components/ui/carousel';
-import { useDateContext } from '@/app/(home)/contexts/date.context';
-import { CALENDAR_STAND_COUNT } from '@/app/(home)/constants/calendar.constant';
-import { getCalendarStartDate } from '@/app/(home)/utils/calendar.util';
-import DayLabel from './day-label';
-import HomeCalendarWeekItem from './home-calendar-week-item';
-import { CarouselWeek } from '../../types/calendar.type';
 import { formatDateWithDash } from '@/utils/format.util';
+import { useDateContext } from '@/app/(home)/_contexts/date.context';
+import { getCalendarStartDate } from '@/app/(home)/_utils/calendar.util';
+import { CALENDAR_STAND_COUNT } from '@/app/(home)/_constants/calendar.constant';
+import DayLabel from './day-label';
+import HomeCalendarMonthItem from './home-calendar-month-item';
+import { CarouselMonth } from '../../_types/calendar.type';
 import { useUserStore } from '@/store/user-store';
-import useCalendarCarousel from '../../hooks/use-calendar-carousel';
+import useCalendarCarousel from '../../_hooks/use-calendar-carousel';
 
-const HomeCalendarWeek = () => {
+const HomeCalendarMonth = () => {
   const { selectedDate, dailyMealCalories } = useDateContext();
   const { personalInfo } = useUserStore((state) => state.user);
-  const { setApi, items: weeks } = useCalendarCarousel<CarouselWeek[]>('week');
+  const { setApi, items: months } = useCalendarCarousel<CarouselMonth[]>('month');
 
   return (
     <div className="space-y-3">
       <DayLabel />
       <Carousel setApi={setApi} opts={{ startIndex: CALENDAR_STAND_COUNT }}>
         <CarouselContent>
-          {weeks.map((week) => {
-            const firstDay = getCalendarStartDate(week.dates);
+          {months.map((month) => {
+            const firstDay = getCalendarStartDate(month.dates);
             return (
               <CarouselItem key={formatDateWithDash(firstDay)}>
-                <HomeCalendarWeekItem
+                <HomeCalendarMonthItem
+                  month={month.dates}
                   selectedDate={selectedDate}
-                  week={week.dates}
                   dailyMealCalories={dailyMealCalories}
                   dailyCaloriesGoal={personalInfo?.dailyCaloriesGoal ?? 0}
                 />
@@ -37,4 +38,4 @@ const HomeCalendarWeek = () => {
     </div>
   );
 };
-export default HomeCalendarWeek;
+export default HomeCalendarMonth;
