+++ src/app/(home)/_components/home-calendar/home-calendar-week.tsx
@@ -1,32 +1,31 @@
-/* eslint-disable react-hooks/exhaustive-deps */
 import { Carousel, CarouselContent, CarouselItem } from '@/components/ui/carousel';
-import { formatDateWithDash } from '@/utils/format.util';
-import { useDateContext } from '@/app/(home)/contexts/date.context';
-import { getCalendarStartDate } from '@/app/(home)/utils/calendar.util';
-import { CALENDAR_STAND_COUNT } from '@/app/(home)/constants/calendar.constant';
+import { useDateContext } from '@/app/(home)/_contexts/date.context';
+import { CALENDAR_STAND_COUNT } from '@/app/(home)/_constants/calendar.constant';
+import { getCalendarStartDate } from '@/app/(home)/_utils/calendar.util';
 import DayLabel from './day-label';
-import HomeCalendarMonthItem from './home-calendar-month-item';
-import { CarouselMonth } from '../../types/calendar.type';
+import HomeCalendarWeekItem from './home-calendar-week-item';
+import { CarouselWeek } from '../../_types/calendar.type';
+import { formatDateWithDash } from '@/utils/format.util';
 import { useUserStore } from '@/store/user-store';
-import useCalendarCarousel from '../../hooks/use-calendar-carousel';
+import useCalendarCarousel from '../../_hooks/use-calendar-carousel';
 
-const HomeCalendarMonth = () => {
+const HomeCalendarWeek = () => {
   const { selectedDate, dailyMealCalories } = useDateContext();
   const { personalInfo } = useUserStore((state) => state.user);
-  const { setApi, items: months } = useCalendarCarousel<CarouselMonth[]>('month');
+  const { setApi, items: weeks } = useCalendarCarousel<CarouselWeek[]>('week');
 
   return (
     <div className="space-y-3">
       <DayLabel />
       <Carousel setApi={setApi} opts={{ startIndex: CALENDAR_STAND_COUNT }}>
         <CarouselContent>
-          {months.map((month) => {
-            const firstDay = getCalendarStartDate(month.dates);
+          {weeks.map((week) => {
+            const firstDay = getCalendarStartDate(week.dates);
             return (
               <CarouselItem key={formatDateWithDash(firstDay)}>
-                <HomeCalendarMonthItem
-                  month={month.dates}
+                <HomeCalendarWeekItem
                   selectedDate={selectedDate}
+                  week={week.dates}
                   dailyMealCalories={dailyMealCalories}
                   dailyCaloriesGoal={personalInfo?.dailyCaloriesGoal ?? 0}
                 />
@@ -38,4 +37,4 @@ const HomeCalendarMonth = () => {
     </div>
   );
 };
-export default HomeCalendarMonth;
+export default HomeCalendarWeek;
