+++ src/app/(home)/_components/home-calendar/index.tsx
@@ -1,7 +1,7 @@
 import ClientOnly from '@/components/commons/client-only';
 import HomeCalendarClientView from './home-calendar-client-view';
 import HomeCalendarServerView from './home-calendar-server-view';
-import { getMonthCalendarDays, getCalendarDateRange } from '../../utils/calendar.util';
+import { getMonthCalendarDays, getCalendarDateRange } from '../../_utils/calendar.util';
 import { getAllMyDailyCalories } from '@/apis/meal.api';
 
 const HomeCalendar = async () => {
