+++ src/app/(home)/_utils/nutrition-diff.util.ts
@@ -1,4 +1,4 @@
-import { MacronutrientDiffItem } from '@/app/(home)/types/diff.type';
+import { MacronutrientDiffItem } from '@/app/(home)/_types/diff.type';
 import { MealNutrition, NutritionGoal } from '@/types/nutrition.type';
 
 /**
