+++ src/app/(home)/page.tsx
@@ -1,7 +1,7 @@
-import DateProvider from '@/app/(home)/contexts/date.context';
-import HomeCalendar from './components/home-calendar';
-import HomeContent from './components/home-content';
-import AppUseSummary from './components/app-use-summary';
+import DateProvider from '@/app/(home)/_contexts/date.context';
+import HomeCalendar from './_components/home-calendar';
+import HomeContent from './_components/home-content';
+import AppUseSummary from './_components/app-use-summary';
 
 const HomePage = async () => {
   return (
